+++ docs/changelog.md
@@ -12,6 +12,7 @@
 - Теперь [getCustomTop](/func?id=getcustomtop) может составить топ по альбомам.
 - Багфикс: `getCustomTop` за период в один день возвращал пустой массив.
 - Багфикс: `readValidArray` - cannot read property, [тикет](https://github.com/Chimildic/goofy/discussions/38)
+- Багфикс: `findAvailablePosition` - cannot read property, [тикет](https://github.com/Chimildic/goofy/discussions/40)
 - Багфиск: `getLastfmRecentTracks` - queryObj is not defined, [тикет](https://github.com/Chimildic/goofy/discussions/39)
 
 ## Версия 1.4.2