+++ docs/changelog.md
@@ -8,6 +8,10 @@
 
 [Перейти к обновленному коду](https://script.google.com/d/1DnC4H7yjqPV2unMZ_nmB-1bDSJT9wQUJ7Wq-ijF4Nc7Fl3qnbT0FkPSr/edit?usp=sharing).
 
+## Версия 1.4.5
+- Теперь [mineTracks](/func?id=minetracks) может искать ключевые слова в названиях альбомов.
+- В `mineTracks` аргумент `playlistCount` переименован в `itemCount`.
+
 ## Версия 1.4.4
 - Новый фильтр [removeUnavailable](/func?id=removeunavailable).
 - `Cache` может читать/писать файлы с расширением `.txt` при явном указании в имени файла.